v1.6.2 - changed user text
This plugin issues the same text when you do /house set and /region set.
It doesn't differentiate between which command is "activated" when you
run the set command, so it issues its text for both when you hit a block
to set a point. I'll look at how I can fix this, but for now I've just
updated the text itself to make it compatible with both /region and
/house.